feat(infra): move lambda definition to service and add to prod
